pyproject.toml
[project]
name = "data-forge-studio"
version = "0.50.0"
description = "Multi-database management tool with modern PySide6 interface"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Lestat2Lioncourt" }
]
license = { text = "MIT" }
keywords = ["database", "sql", "query-manager", "data-transformation", "gui", "pyside6", "qt"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.10",
    "Topic :: Database :: Front-Ends",
]

dependencies = [
    "pyside6>=6.10.1",
    "colorama>=0.4.6",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pyodbc>=5.3.0",
    "sqlalchemy>=2.0.44",
    "sqlparse>=0.5.4",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
dataforge = "dataforge_studio.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataforge_studio"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
